Module Stocks

module Cli : sig ... end

Module providing functions for user interactions and financial data retrieval.

module Data : sig ... end

Module providing functions for retrieving financial data and information.

module User : sig ... end

Module defines User and functions to interact with a financial system